[BUG FIX] Stop depending on PHP local server behavior
The PHP builtin server serves up the first index.php file it can find
when a file match is not found. Apache does not do this so we need to not
navigate to a page that does not exist. Thus we need to use the moostaka
object's navigate method rather than window.location = ... to avoid a
page load.