Added user entity, user model and unit tests.
In order to make progress, a more flexible role system was needed. In order to manage the complexity of a flexible role system a more disciplined approach to data was needed. This is a start on that path. It provides a rather abstract User type, a model (bridge to the database) and a database model aware or concrete implementation of the user type along with unit tests.  Building out the Role type, model and database aware implementation should allow for a rewrite of the current Security.php that exposes fine grained permissions.