Begin using entity abstraction in API endpoints
Switched over to using new entity abstraction layer for API endpoints. Keeping the script for each endpoint design because it is simple and still works. Only /api/config.php, /api/login.php and /api/equipment[GET] are updated but the design is new design is working and we have permissions to the front end now.